update dashboard cqrs

+++ src/core/modules/project_management/infrastructure/reporting/models/report_models.py
@@ -49,11 +49,11 @@ class ProjectKPI:
     tasks_not_started: int
     critical_tasks: int
     late_tasks: int
-    total_planned_cost: float | None
-    total_actual_cost: float | None
-    cost_variance: float | None
-    total_committed_cost: float | None
-    committment_variance: float | None
+    total_planned_cost: Decimal | None
+    total_actual_cost: Decimal | None
+    cost_variance: Decimal | None
+    total_committed_cost: Decimal | None
+    committment_variance: Decimal | None
     financial_detail_included: bool = True
 
 @dataclass
